chore: extract helper functions into separate module, extract default options, separate some of the core logic, update dependencies, use new ESLint configuration
