+++ .eslintrc
@@ -1,3 +1,3 @@
 {
-  "extends": "@dtrussia/eslint-config-dtrussia/react"
+  "extends": "@dtrussia/eslint-config-dtrussia"
 }
